LayoutTests/imported/w3c/web-platform-tests/storage-access-api/storage-access-beyond-cookies.indexedDB.sub.https.window-expected.txt


FAIL Verify StorageAccessAPIBeyondCookies for IndexedDB assert_equals: Storage Access API should be accessible and return first-party data expected "HasAccess for indexedDB" but got "Unable to load handle in cross-site context for all"
